Adds further configuration
